cognificar_leiloes/requirements.txt
pandas
numpy
selenium-wire
beautifulsoup4
google
webdriver-manager
psutil
azure-storage-blob